requirements.txt
openai
requests
weaviate-client
pdfplumber
